test(release-matrix): split test_release_matrix.py into tests/release_matrix/ package
The single-file suite grew to ~970 lines, hard to navigate for a proposal
meant to be reviewed line-by-line. Split by functional axis into
tests/release_matrix/: specs.py (Spec dataclass, config constants, the
SPECS/EP2_SPECS/TP2_SPECS/TPEP_SPECS/MULTIRANK_SPECS matrix), docker.py
(container lifecycle: build_serve_cmd/build_ep2_serve_cmd, start/wait/stop,
warmup, EP=2 readiness), api.py (HTTP-API helpers: chat, text extraction,
tps calculation, repetition-loop detection, long-context/tool fixtures),
assertions.py (the five shared assertion bodies), and baselines.py
(baseline load/write helpers). tests/test_release_matrix.py now only holds
the deployed_atlas/deployed_atlas_ep2 fixtures and the two test classes,
importing everything else from release_matrix.*.

No behavior change: same 21 specs, same 5 test categories per spec, same
105 collected tests, same -k point-debug and -m multirank isolation.
BASELINE_DIR now resolves via the package's parent directory instead of
__file__'s own directory, since it moved one level deeper.

tests/conftest.py is untouched — it only holds the two pytest hooks
(--update-baselines option, multirank marker) and doesn't need splitting.
